Revert "Merge branch 'master' of https://github.com/hoangkimkhai276/Pentominoes_Phase_3"
This reverts commit 2e9ab2ea1d51211caff03d00ae44df0b8b4186aa, reversing
changes made to 290811b0cf0d45a640b661949d4947b1d8e80e93.

+++ knapsack/parcel/Parcel.java
@@ -38,7 +38,7 @@ public interface Parcel {
 	int getVolume();
 	
 	default double getDensity() {
-		return getValue() / ((double)getVolume());
+		return getValue() / getVolume();
 	}
 
 	Point3D[] getPoints();
